requirement.txt
pickle
appdirs==1.4.3
cycler==0.10.0
funcsigs==1.0.2
functools32==3.2.3.post2
matplotlib==2.0.2
mecab-python==0.996
mock==2.0.0
mojimoji==0.0.5
nltk==3.2.4
numpy==1.12.1
packaging==16.8
pandas==0.20.1
pbr==3.0.0
protobuf==3.3.0
pyparsing==2.2.0
pyrouge==0.1.3
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scipy==0.19.0
seaborn==0.7.1
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.1.0
Werkzeug==0.12.1
